### 2025-01-30 7b08106
drop fields from paper.md/yaml header
