+++ paper/paper.md
@@ -179,12 +179,6 @@ When merging the contributions, we include students as contributors of the packa
 
 # Teaching materials
 
-<!--
-Options:
-- instructional design
-- teaching experience
--->
-
 The delivery of the course is designed to foster active learning 
 through a blend of in-person sessions, individual group work, and interactive hacking sessions, 
 all facilitated by the instructor.
@@ -215,62 +209,7 @@ while emphasizing practical applications that enhance both technical skills and 
 \end{tabular}
 \end{table}
 
-<!-- | Resource                                                                                          | Description and focus                                                                                                                                                |
-|---------------------------------------------------------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------|
-
-| [Landing page](https://digital-work-lab.github.io/open-source-project/) | 
-Provides an accessible overview of the course, aimed to engage students. |
-| [Syllabus](https://digital-work-lab.github.io/open-source-project/docs/syllabus.html) | 
-Offers a structured, detailed overview of course objectives, content, and pedagogical approach, 
-complementing the landing page with broader course context. |
-| [Slides](https://digital-work-lab.github.io/open-source-project/docs/syllabus.html#materials) | 
-Slides in Markdown (Marp) format. |
-| [Notebooks](https://digital-work-lab.github.io/open-source-project/docs/syllabus.html#materials) | 
-Designed to engage directly with the Python package, using `git reset` to access solutions, 
-supporting a smooth, practical learning flow. |
-| [Teaching notes](https://digital-work-lab.github.io/open-source-project/docs/teaching_notes.html) | 
-The teaching notes contain preparation checklists, scheduled mailings, session readers, and a concept. |
-
-Table: Materials -->
-
 # Pedagogical considerations
-
-<!-- 
-Teaching Python package development requires a structured approach that balances simplicity with depth, 
-ensuring students build a solid foundation before progressing to more advanced topics.
-The course design is informed by key pedagogical principles aimed at reducing cognitive load, 
-promoting gradual learning progression, and encouraging active participation.
-Below, we outline three core principles that shape the design of our educational resources.
-
-1. **Select and Simplify**
-
-To reduce cognitive overload and potential errors, we prioritize simplicity in both tools and workflows.
-For instance, we use GitHub Codespaces as the development environment to standardize setups across all learners, 
-thereby eliminating issues related to different operating systems and environment configurations.
-Additionally, we streamline the package management system by focusing on poetry, 
-rather than offering multiple options like *hatch*, *PDM*, *uv*, or *pipenv*.
-This focused approach aligns with cognitive load theory [@Sweller1988], 
-helping students concentrate on core concepts without unnecessary distractions.
-
-2. **Gradually Progress in Complexity**
-
-The course is structured to ensure students gradually progress from foundational to more complex topics.
-Starting with basic Python and Git skills, the material builds incrementally, 
-introducing concepts like dependency management and package distribution only after learners have mastered simpler tasks.
-This approach ensures that students develop a robust understanding of each topic before moving on, 
-reducing the risk of overwhelming them [@Anderson2001].
-
-3. **Learn interactively and in groups**
-
-Interactive and collaborative learning plays a crucial role in student engagement and knowledge retention [@Guzdial1998].
-The course incorporates live coding sessions, and group-based exercises to make practices as accessible as possible 
-and encourage active participation [@VialNegoita2018].
-We build on the principles of active learning to promote deeper understanding through hands-on practice and peer collaboration.
-In particular, group problem-solving can foster a collaborative environment where learners can exchange ideas, 
[…]
