### 2025-01-30 #275 @geritwagner
Thank you for opening the pre-review issue. I just fixed the `paper.md` to ensure that the pdf is generated successfully and removed material in comments for a more adequate word count. I also included the standard Creative Commons (CC-BY) license.
