### 2025-01-30 a2b3842
Update week_3_resources.md
