### 2025-01-30 6492ee5
Update week_3_python_packages.md
