### 2025-01-30 5911cab
Update syllabus.md
