### 2025-01-30 b94b44c
Update teaching_notes.md
