### 2025-01-30 90b184d
Update week_3_python_notebook_2.md
