### 2025-01-30 d7538aa
Update week_3_python_notebook_2.md
