+++ .github/workflows/labot.yml
@@ -8,6 +8,8 @@ on:
   pull_request:
     branches:
       - main
+  issue_comment:
+    types: [created]
 
 jobs:
   install-and-run:
