### 2025-01-31 de30966
Update week_3_python_notebook_2.md
