+++ .github/workflows/labot.yml
@@ -8,8 +8,6 @@ on:
   pull_request:
     branches:
       - main
-  issue_comment:
-    types: [created]
 
 jobs:
   install-and-run:
