### 2025-03-23 86df9bd
Update week_1_teaching_notes.md
