### 2025-03-27 94b9048
Update syllabus.md
