### 2025-03-27 42cecc1
revise criteria
