### 2025-03-27 9496962
Update syllabus.md
