+++ index.md
@@ -80,17 +80,17 @@ The outline for sessions is below:
 | 2    | [Introduction to Git](docs/week_2_git.html)          | May 2, 08.00-12.00                                    | WE5 01.003                |
 | 3    | [Introduction to Python](docs/week_3_python.html)    | May 8, 12.00-14.00  <br> May 8, 14.00-16.00           | WE5 01.004 <br> WE5 01.003|
 | 5    | [Best practices](docs/week_5_best_practices.html)    | May 9, 08.00-12.00                                    | WE5 01.003                |
-| 6-11 | [Group work](docs/week_6-11_group_work.html) with hacking sessions | On demand                               | tba                       |
+| 6-11 | [Group work](docs/week_6-11_group_work.html) with hacking sessions | On demand                               | WE5 01.081                |
 | 12   | [Code review session](docs/week_12_code_review.html) | May 22, 12.00-14.00                                   | WE5 01.004                |
 
 ## Deliverables
 
 | Deliverable                                                                           | Date            |
 | ------------------------------------------------------------------------------------- | --------------- |
-| Create a GitHub account and contribute to issue discussions to form teams             | tba             |
-| Contribute code to a fork, create branches and merge code                             | tba             |
-| Create the pull request (code is evaluated on January 12)                             | tba             |
-| Implement the feedback provided in the pull request                                   | tba             |
+| Create a GitHub account and contribute to issue discussions to form teams             | End of week 1   |
+| Contribute code to a fork, create branches, and merge code                            | End of week 6   |
+| Create the pull request (code is evaluated in week 12)                                | End of week 10  |
+| Implement the feedback provided in the pull request                                   | End of week 13  |
 
 ## Instructor
 
