+++ index.md
@@ -76,12 +76,12 @@ The outline for the sessions is below:
 
 | Week | Session                                              | Date and Time                                         | Location                  | Status |
 | ---- | ---------------------------------------------------- | ----------------------------------------------------- | ------------------------- | ------ |
-| 1    | [Finding the topic](docs/week_1_topic.html)          | April 23, 12.00-14.00                                 | WE5 04.003                |{% assign d='2025-03-15' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
-| 2    | [Introduction to Git](docs/week_2_git.html)          | May 2, 08.00-12.00                                    | WE5 01.003                |{% assign d='2025-04-15' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
-| 3    | [Introduction to Python](docs/week_3_python.html)    | May 8, 12.00-14.00  <br> May 8, 14.00-16.00           | WE5 01.004 <br> WE5 01.003|        |
-| 5    | [Best practices](docs/week_5_best_practices.html)    | May 9, 08.00-12.00                                    | WE5 01.003                |        |
-| 6-11 | [Group work](docs/week_6-11_group_work.html) with hacking sessions | On demand                               | WE5 01.081                |        |
-| 12   | [Code review session](docs/week_12_code_review.html) | May 22, 12.00-14.00                                   | WE5 01.004                |        |
+| 1    | [Finding the topic](docs/week_1_topic.html)          | April 23, 12.00-14.00                                 | WE5 04.003                |{% assign d='2025-04-23' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
+| 2    | [Introduction to Git](docs/week_2_git.html)          | May 2, 08.00-12.00                                    | WE5 01.003                |{% assign d='2025-05-02' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
+| 3    | [Introduction to Python](docs/week_3_python.html)    | May 8, 12.00-16.00  <br> May 9, 08.00-12.00           | WE5 01.004 <br> WE5 01.003 <br> WE5 01.003|{% assign d='2025-05-08' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
+| 5    | [Best practices](docs/week_5_best_practices.html)    | May 22, 12.00-14.00                                   | WE5 01.004                |{% assign d='2025-05-09' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
+| 6-11 | [Group work](docs/week_6-11_group_work.html) with hacking sessions | On demand                               | WE5 01.081                |{% assign d='2025-05-22' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
+| 12   | [Code review session](docs/week_12_code_review.html) | TBA                                                   | TBA                       |{% assign d='2025-07-10' | date: "%s" %}{% assign now='now' | date: "%s" %}{% if d <= now %}🟢 Completed{% else %}⚪ Upcoming{% endif %} |
 
 ## Deliverables
 
