### 2025-04-22 84f5dd4
update first session slides
