+++ slides/01-topics.md
@@ -57,10 +57,13 @@ Otto-Friedrich-Universität Bamberg
 
 # What you will learn (III): Open-Sourcing
 
-- Work with the Open-Source workflows of GitHub
-- Make a first contribution to a public Open-Source project
-- Adopt the role of a maintainer and review the code of your peers
-- Have an opportunity to create your *developer portfolio*
+- Open-source software plays a major role in the industry (see [HBS working paper](https://www.hbs.edu/faculty/Pages/item.aspx?num=65230)) and firms start to use open-source principles to organize their work (see [innersource](https://github.com/resources/articles/software-development/innersource))
+- You will
+
+  - Work with the Open-Source workflows of GitHub
+  - Make a first contribution to a public Open-Source project
+  - Adopt the role of a maintainer and review the code of your peers
+  - Have an opportunity to create your *developer portfolio*
 
 ![bg right:35% width:350px](../assets/github-contributors.png)
 
@@ -99,7 +102,7 @@ Dates and rooms are available [online](https://digital-work-lab.github.io/open-s
 # How groups will form
 
 You **assign yourself to a group** when you contribute code, following these steps:
-1. **Signal your intent** to contribute by joining the issue discussion on GitHub (in the [issue discussion](https://github.com/CoLRev-Environment/colrev/milestones), before the **Git Session**, for two topics at most).
+1. **Signal your intent** to contribute by joining the issue discussion on GitHub (in the [issue discussion](https://github.com/CoLRev-Environment/colrev/issues?q=is%3Aissue%20state%3Aopen%20label%3A%22good%20first%20issue%22%20%20label%3Asearch_source), before the **Git Session**, for two topics at most).
 2. **Organize your work**, meet potential group members, and select a project leader (in the following sessions). Keep in mind that **no more than 5 people** will be accepted per group.
 3. **Officially sign up for a group** by contributing a non-trivial code part (\*). Team members' contributions must be made in separate commits. Select a project lead who sends your GitHub-ID together with a link to your code contribution, your student ID, and e-mail address to [gerit.wagner@uni-bamberg.de](mailto:gerit.wagner@uni-bamberg.de).
 
@@ -111,7 +114,7 @@ This process ensures that **all group members contribute fairly**, as typically,
 
 # Evaluation criteria
 
-**Process: Open-Source Practices**
+**Process: Open-Source practices**
 
 - Follow established Git conventions for commit messages, branching, and pull requests.
 - Actively engage in the community by reviewing code, providing and incorporating constructive feedback.
@@ -134,15 +137,15 @@ This process ensures that **all group members contribute fairly**, as typically,
 
 Evaluation is contingent on completing the deliverables listed [here](https://digital-work-lab.github.io/open-source-project/#deliverables).
 
-## Evaluation Components
+**Evaluation components**
 
-- **Code Review Session (30%)**  
+- **Code review session (30%)**  
   Focus on collaboration, version control practices, and understanding of code.
 
-- **Project Code (70%)**  
+- **Project code (70%)**  
   Assessment of functionality, overall code quality, and implementation of feedback.
 
-## Not Considered in the Evaluation
+**Not considered in the evaluation**
 
 - Training sessions (Topics, Git, Python)
 - Hacking sessions
@@ -161,7 +164,7 @@ Evaluation is contingent on completing the deliverables listed [here](https://di
 
 ---
 
-# Finding things in an Open Source project
+# Finding things in an open-source project
 
 Go to the [CoLRev project repository](https://github.com/CoLRev-Ecosystem/colrev).
 
@@ -196,7 +199,7 @@ blockquote {
 
 **Find a topic**
 
[…]
