### 2025-04-22 a174432
reset semester
