+++ docs/teaching_notes.md
@@ -24,13 +24,13 @@ Based on the [pedagogical concept](pedagogy.html). Administrative tasks are [her
 
 ## Preparation
 
-- [x] Announce project on [instagram](https://www.instagram.com/informatik_unibamberg/){: target="_blank"}
-- [x] Advertise the project in the lecture or other courses based on the [mailing](mailings.html#advertising) - check if not applicable
+- [ ] Announce project on [instagram](https://www.instagram.com/informatik_unibamberg/){: target="_blank"}
+- [ ] Advertise the project in the lecture or other courses based on the [mailing](mailings.html#advertising) - check if not applicable
 - [x] Update dates and rooms for sessions
 - [x] Update the rating average in the badge (change the badge-link on index.md) and participation
-- [x] Review the [pedagogical concept](pedagogy.html) in preparation of the project
-- [x] Check whether `pre-commit run --all` runs without errors in Codespaces
-- [x] Check whether `colrev package --init` works
+- [ ] Review the [pedagogical concept](pedagogy.html) in preparation of the project
+- [ ] Check whether `pre-commit run --all` runs without errors in Codespaces
+- [ ] Check whether `colrev package --init` works
 
 <details markdown="block">
 <summary>Student communication notes</summary>
@@ -50,18 +50,18 @@ Based on the [pedagogical concept](pedagogy.html). Administrative tasks are [her
 
 ## Week 1: Topics <a id="topics"></a>
 
-- [x] Send out [welcome and information mailing](mailings.html#welcome) before the first session
-- [x] Link the [feedback issue](https://github.com/digital-work-lab/open-source-project/issues){: target="_blank"} in VC, explain our process of evaluation and improvement, summarize our improvements from last semester
+- [ ] Send out [welcome and information mailing](mailings.html#welcome) before the first session
+- [ ] Link the [feedback issue](https://github.com/digital-work-lab/open-source-project/issues){: target="_blank"} in VC, explain our process of evaluation and improvement, summarize our improvements from last semester
 
 ## Week 2: Git  <a id="git"></a>
 
-- [x] Check whether the `rec_dict.patch` applies (`git diff > rec_dict.patch`)
-- [x] Print some of the [overviews](https://github.com/digital-work-lab/practice-git/blob/main/notebooks/img/overview-task.pdf){: target="_blank"}
-- [x] Send out [comment-on-issues mailing](mailings.html#comment-issues)
+- [ ] Check whether the `rec_dict.patch` applies (`git diff > rec_dict.patch`)
+- [ ] Print some of the [overviews](https://github.com/digital-work-lab/practice-git/blob/main/notebooks/img/overview-task.pdf){: target="_blank"}
+- [ ] Send out [comment-on-issues mailing](mailings.html#comment-issues)
 
 ## Week 3: Python <a id="python"></a>
 
-- [x] Prepare the `tutorial_python` branch and update the commit-SHAs in the notebooks
+- [ ] Prepare the `tutorial_python` branch and update the commit-SHAs in the notebooks
 
 <details markdown="block">
 <summary>Python prep</summary>
@@ -78,7 +78,7 @@ git push -f
 When the poetry.lock/pyproject.toml fail: checkout --ours poetry.lock/pyproject.toml, poetry add bibtexparser
 </details>
 
-- [x] Create a list of topics and students (first/second choice) and facilitate issue discussions
+- [ ] Create a list of topics and students (first/second choice) and facilitate issue discussions
 
 <details markdown="block">
 <summary>Facilitating issue discussions</summary>
@@ -89,13 +89,13 @@ When the poetry.lock/pyproject.toml fail: checkout --ours poetry.lock/pyproject.
 
 ## Week 4: Group work (no session)
 
[…]
