+++ .wordlist.txt
@@ -979,3 +979,6 @@ tba
 endfor
 endif
 OSP
+HBS
+innersource
+abd
