### 2025-04-29 #275 @labarba
@NickleDave — would you be willing to help us out with a review of this submission to The Journal of Open Source Education, JOSE? We still need to assign a handling editor, but I thought of you for reviewing since you did such a wonderful job of presenting about Python packaging at last year's [URSSI Summer School](https://github.com/si2-urssi/summerschool-June2024) at GW.

Might you also suggest a couple of other people to review? (We need two to proceed.)
