+++ docs/syllabus.md
@@ -122,7 +122,6 @@ We reserve the right to assign different grades within the group and give extra 
 
 - Follow established Git conventions for commit messages, branching, and pull requests.
 - Actively engage in the community by reviewing code, providing and incorporating constructive feedback.
-- Contribute professionally to project discussions, issues, and decision-making.
 
 **Code: Functionality**
 
