### 2025-04-30 #275 @NickleDave
Hi @labarba I'm happy to review.

I see the intro references pyOpenSci and py-pkgs.org -- I wonder if @lwasser @tomasbeuzen @ttimbers have ideas for potential reviewers (apologies for tagging you in, I know y'all are busy folks)
