+++ slides/02-git.md
@@ -127,7 +127,7 @@ blockquote {
 
 - Commits form a **directed acyclic Graph (DAG)**, i.e., all commits can have one or more children, and one or more parents (except for the first commit, which has no parent). Closed directed cycles are not allowed.
 - With the **git branch \<branch-name\>** command, a separate line of commits can be started, i.e., one where different lines of commits are developed from the same parent. The branch pointer typically points at the latest commit in the line.
-- With the **git switch \<branch-name\>** command, we can select the branch on which we want to work. Switch effectively moves the HEAD pointer, which points to a particular branch and indicates where new commits are added.
+- With the **git switch \<branch-name\>** command, we can select the branch on which we want to work. Switch effectively moves the **HEAD** pointer, which points to a particular branch and indicates where new commits are added.
 - With the **git merge \<other-branch\>** command, separate lines of commits can be brought together, i.e., creating a commit with two parents. The *merge commit* integrates the contents from the *\<other-branch\>* into the branch that is currently selected. The *\<other-branch\>* is not changed. 
 - Per default, Git sets up a branch named "main".
 
