+++ _config.yml
@@ -12,6 +12,8 @@ gh_edit_repository: 'https://github.com/digital-work-lab/open-source-project'
 gh_edit_branch: 'main'
 gh_edit_view_mode: 'tree'
 
+enable_copy_code_button: true
+
 mermaid:
   version: "11.4.0"
 

[…]
