+++ .wordlist.txt
@@ -982,3 +982,4 @@ OSP
 HBS
 innersource
 abd
+uv
