+++ docs/week_2_git_notebook_merge_conflict.md
@@ -23,13 +23,13 @@ We  <img src="../assets/iconmonstr-favorite-2.svg" alt="Edit" width="12" height=
 
 With this notebook, you can practice merging and resolving merge conflicts.
 
-| Practice | Label                                                   | Time (min) |
-|----------|---------------------------------------------------------|------------|
-|  1       | [A simple merge conflict](#simple)                      | 20         |
-|  2       | Optional: [A more realistic merge conflict](#realistic) | 30         |
-|  3       | [Understand how to prevent merge conflicts](#prevent)   |  8         |
-|  4       | [Wrap-up](#wrap-up)                                     |  2         |
-|          | Overall                                                 | 40         |
+| Practice | Label                                                    | Time (min) |
+|----------|----------------------------------------------------------|------------|
+|  1       | [A simple merge conflict](#simple)                       | 20         |
+|  2       | [A more realistic merge conflict](#realistic) (Optional) | 30         |
+|  3       | [Understand how to prevent merge conflicts](#prevent)    |  8         |
+|  4       | [Wrap-up](#wrap-up)                                      |  2         |
+|          | Overall                                                  | 40         |
 
 <img src="../assets/iconmonstr-certificate-6.svg" alt="Edit" width="12" height="12"> We are here to help if errors or questions come up!
 
@@ -115,7 +115,6 @@ The first section shows the content on your current branch (selected by `HEAD`).
 
 **Task**: To resolve the conflict, select the code parts that should be retained. Remove the surrounding conflict markers. Once completed, complete the merge.
 
-
 ```python
 git add app.py
 git commit
@@ -131,7 +130,6 @@ Check the Git history and verify whether the merge commit was created.
 
 For this task, we focus on a project where Lisa and Ted work on a Python module for a command-line application. Lisa works on a separate branch and changes `print()`statements to `logging.info()`. In parallel, Ted introduces changes to the `COLORS` variable on main. Both changes affect the same module and even the same lines of code. Let's see how the conflict unfolds and how we can resolve it.
 
-
 ```python
 # Run these commands if you do not have the colrev project.
 # git clone https://github.com/CoLRev-Environment/colrev
@@ -139,7 +137,6 @@ For this task, we focus on a project where Lisa and Ted work on a Python module 
 # code -a /workspaces/colrev
 # code /workspaces/practice-git/notebooks/git_merge_conflicts.ipynb
 ```
-
 
 ```python
 cd /workspaces/colrev
@@ -157,7 +154,6 @@ git apply /workspaces/practice-git/notebooks/use-logger.patch
 - Analyze the changes using the Git GUI
 - Commit the changes using an informative commit message
 
-
 ```python
 # We go back to the main branch and apply another patch
 git switch main
@@ -171,8 +167,6 @@ git apply /workspaces/practice-git/notebooks/replace-color-import.patch
 - Commit the changes
 - Merge the logger branch and resolve the conflict, making sure that the changes of Lisa and Ted are retained correctly
 
-
-
 ```python
 git merge logger
 # Resolve merge conflict
@@ -181,7 +175,6 @@ git status
 ```
 
 To validate your solution, run the following command to retrieve our solution:
-
 
 ```python
 cp -f /workspaces/practice-git/cli.py colrev/ui_cli/cli.py
@@ -193,7 +186,6 @@ Note: for simplicity, we worked in the same repository. However, the merge confl
 ## Part 3: Understand how to prevent merge conflicts <a id="prevent"></a>
 
 Although merge conflicts are not necessarily a bad thing, it is recommended to avoid complex merge conflicts. To achieve this, different strategies can be combined:
-
 
 - Understand whether changes are likely to create merge conflicts
[…]
