### 2025-05-02 d894f19
update notes etcs
