+++ docs/week_2_git_teaching_notes.md
@@ -10,21 +10,17 @@ search_exclude: true
 
 | Time (start)   | Duration  | Topic                                     | Additional materials                                                                                                                                |
 |----------------|-----------|-------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------|
-| 00:00          | 90        | [Branching](#branching)                   | Concepts + Hands-on practice ([branching notebook](https://github.com/digital-work-lab/practice-git?tab=readme-ov-file){: target="_blank"})         |
+| 00:00          | 90        | [Branching](#branching)                   | Concepts + Hands-on practice ([branching notebook](week_2_git_notebook_branching){: target="_blank"})         |
 | 01:30          | 5         | Break                                     |                                                                                                                                                     |
-| 01:35          | 40        | [Committing](#committing)                 | Concepts + Hands-on practice ([committing notebook](https://github.com/digital-work-lab/practice-git?tab=readme-ov-file){: target="_blank"})        |
+| 01:35          | 40        | [Committing](#committing)                 | Concepts + Hands-on practice ([committing notebook](week_2_git_notebook_committing){: target="_blank"})        |
 | 02:15          | 5         | Break                                     |                                                                                                                                                     |
-| 02:20          | 25        | [Transfer: Conflicts](#transfer-conflicts)| Conflict resolution task ([conflict-resolution notebook](https://github.com/digital-work-lab/practice-git?tab=readme-ov-file){: target="_blank"})   |
+| 02:20          | 25        | [Transfer: Conflicts](#transfer-conflicts)| Conflict resolution task ([conflict-resolution notebook](week_2_git_notebook_merge_conflict){: target="_blank"})   |
 | 02:45          | 5         | Break                                     |                                                                                                                                                     |
-| 02:50          | 15        | [Collaborating](#collaborating)           | Concepts + Group practice (do at home) ([collaborating notebook](https://github.com/digital-work-lab/practice-git?tab=readme-ov-file){: target="_blank"})        |
+| 02:50          | 15        | [Collaborating](#collaborating)           | Concepts + Group practice (do at home) ([collaborating notebook](week_2_git_notebook_collaboration){: target="_blank"})        |
 | 03:00          | 10        | [Wrap-up](#wrap-up)                       |                                                                                                                                                     |
 
-{: .highlight }
+{: .info }
 > Carefully prepare the explanation of the "three states of a file" (focusing on the **states of a file** and explaining the commands and state changes based on the workflow).
-
-{: .objective }
-> - **Goal**: figure out how to accomplish the tasks (the instructions are more detailed at the beginning, you need to remember the commands/use and annotate the git cheatsheet. You will have to use the commands again, know what they do, and how the changes and commands are situated in the three areas)
-> - We work with the CLI.
 
 <div class="page-break"></div>
 
@@ -43,8 +39,12 @@ Check whether students are forming groups, whether there are any challenges.
 Today, we want to understand how to use Git for software development in teams.
 This session is divided into three parts.
 
-- We’ll start with **branching**, then move on to **committing**, and finally, we'll look at **collaborating**.
-- Each of these parts includes **concepts** that will be discussed, along with **related exercises**.
+{: .objective }
+>- We will start with **branching**, then move on to **committing**, and finally, we'll look at **collaborating**.
+> - Each of these parts includes **concepts** that will be discussed, along with **related exercises**.
+
[…]
