### 2025-05-02 66ae8be
update first part of teaching note
