+++ slides/03-python_1.md
@@ -120,5 +120,8 @@ git reset --hard ca9902e666518af1d33a368adf055c9809004433
 - Read the [package development](https://colrev-environment.github.io/colrev/dev_docs/packages.html) documentation.
 - Study code of related CoLRev packages.
 - Take notes on the CoLRev-objects or libraries that will be needed.
+- Remember to **stop the codespace** ([Codespaces overview](https://github.com/codespaces))
+
+![bg right:40% width:400px](../assets/codespace-stop.png)
 
 Note: These [tutorials](https://www.codecademy.com/catalog/language/python?g_network=g&g_productchannel=&g_adid=624888211335&g_locinterest=&g_keyword=codecademy%27s%20learn%20python&g_acctid=243-039-7011&g_adtype=&g_keywordid=kwd-2259230975260&g_ifcreative=&g_campaign=account&g_locphysical=9042755&g_adgroupid=128133971748&g_productid=&g_source={sourceid}&g_merchantid=&g_placement=&g_partition=&g_campaignid=12575778360&g_ifproduct=&utm_id=t_kwd-2259230975260:ag_128133971748:cp_12575778360:n_g:d_c&utm_source=google&utm_medium=paid-search&utm_term=codecademy%27s%20learn%20python&utm_campaign=INTL_Brand_Phrase&utm_content=624888211335&g_adtype=search&g_acctid=243-039-7011&gad_source=1&gclid=CjwKCAjwzIK1BhAuEiwAHQmU3t-FhdTeqJcFCUfaU1x1cvYdKGfh9vdxz1b8ll5Tl_2wcSc8wgMNhRoCJqoQAvD_BwE) may be helpful to continue practicing Python
