### 2025-05-22 c6bec20
update teaching notes
