### 2025-05-26 f2d5cbf
update notes
