+++ .github/workflows/links_fix.yml
@@ -21,13 +21,16 @@ jobs:
         id: lychee
         uses: lycheeverse/lychee-action@v1
 
-      - name: Create Issue From File
+      - name: Create or Update Link Checker Issue
         if: env.lychee_exit_code != 0
         uses: peter-evans/create-issue-from-file@v5
         with:
+          issue-number: 228
           title: Link Checker Report
           content-filepath: ./lychee/out.md
           labels: report, automated issue
+          reopen-closed: true
+
   internalLinkChecker:
     runs-on: ubuntu-latest
     steps:
