### 2025-06-17 #275 @labarba
👋  @ashleefv – could you take this JOSE submission as editor? ([Teaching Python package development: A structured course with learning resources and an instructor's guide](https://github.com/openjournals/jose-reviews/issues/275#top))
We already have one agreed reviewer, and perhaps a second reviewer could be one of the authors of this other JOSE paper:
https://jose.theoj.org/papers/10.21105/jose.00130
