### 2025-07-08 14ff72c
Fix typos in docs and metadata
