### 2025-07-08 4feba1d
Fix typos in docs and metadata
