### 2025-07-10 #275 @ashleefv
@ACharbonneau @abhijna @marisalim @s-canchi @ctb Are you available to review this submission to The Journal of Open Source Education, JOSE? It seems related to your own previous publication: https://jose.theoj.org/papers/10.21105/jose.00130
