### 2025-07-10 #275 @ctb
I'm interested, but only available in 3 weeks (first week of August).
