### 2025-07-10 #292 @editorialbot
[REVIEW]: Teaching Python package development: A structured course with learning resources and an instructor's guide
**Submitting author:** @geritwagner (<a href="http://orcid.org/0000-0003-3926-7717">Gerit Wagner</a>)
**Repository:** https://github.com/digital-work-lab/open-source-project
**Branch with paper.md** (empty if default branch): jose
**Version:** v0.4.0
