### 2025-07-10 #292 @geritwagner
Thank you for opening the issue and helping with the review of our work! I have just updated the references to fix the two DOI errors (missing/invalid).
