### 2025-07-10 #292 @NickleDave
- Issue: https://github.com/digital-work-lab/open-source-project/issues/238

    - Issue: https://github.com/digital-work-lab/open-source-project/issues/239

I feel that the course as currently written is not quite substantial and appropriate for the target audience.
Issue discussing this here: https://github.com/digital-work-lab/open-source-project/issues/241

    - Issue: https://github.com/digital-work-lab/open-source-project/issues/236
    - Issue: https://github.com/digital-work-lab/open-source-project/issues/237
