+++ README.md
@@ -10,6 +10,7 @@
 ![Documentation Status](https://img.shields.io/github/actions/workflow/status/digital-work-lab/open-source-project/pages.yml?label=documentation)
 ![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
 [![DOI](https://zenodo.org/badge/697812022.svg)](https://doi.org/10.5281/zenodo.13871275)
+[![status](https://jose.theoj.org/papers/cc49b9604b9fd10bc8524d1539620be6/status.svg)](https://jose.theoj.org/papers/cc49b9604b9fd10bc8524d1539620be6)
 
 </div>
 
