### 2025-07-20 a47655d
Update syllabus.md
