+++ index.md
@@ -36,7 +36,7 @@ permalink: /
 <!-- ![Offered by: Digital Work at Otto-Friedrich-Universität Bamberg](https://img.shields.io/badge/Offered%20by-%20Digital%20Work%20(Otto--Friedrich--Universit%C3%A4t%20Bamberg)-blue) -->
 [![Evaluations](https://img.shields.io/badge/Rating-★★★★★%20({{ average_of_averages_rounded }}%20/%205)-yellow)](https://digital-work-lab.github.io/open-source-project/docs/evaluations.html) 
 ![Completion](https://img.shields.io/badge/Enrollment-{{ total_students }}%20students-green)
-[![Peer-Reviewed Quality](https://img.shields.io/badge/Pedagogical%20Foundations-Peer%20Reviewed-green)](https://digital-work-lab.github.io/rethink-git-teaching/){: target="_blank"}
+[![Peer-Reviewed Quality](https://img.shields.io/badge/Pedagogical%20Foundations-Peer%20Reviewed-green)](https://digital-work-lab.github.io/open-source-project/docs/syllabus.html#teaching-approach){: target="_blank"}
 ![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
 
 Welcome!
