### 2025-07-20 17c0a28
Update syllabus.md
