### 2025-07-20 ba96d52
Update syllabus.md
