### 2025-07-20 7769df9
Update syllabus.md
