### 2025-07-22 0c3c326
Update week_12_code_review_teaching_notes.md
