+++ .wordlist.txt
@@ -982,3 +982,13 @@ HBS
 innersource
 abd
 uv
+CAIS
+FBTKM
+JISE
+aisnet
+bc
+cais
+fd
+jise
+jose
+theoj
