### 2025-08-01 fc1bdd7
Update teaching_notes.md
