### 2025-08-01 f7274fb
Update teaching_notes.md
