### 2025-08-01 43dc237
Update teaching_notes.md
