### 2025-08-04 #292 @NickleDave
Thank you @ashleefv -- the authors contacted me by email and I told them I was still working through the checklist, and will raise issues on their repo. I haven't gotten to that yet but will have it done by this coming Monday
