### 2025-08-11 digital-work-lab/open-source-project#236 @NickleDave
JOSE[paper]: Describe how this class has been used in the classroom or other settings
Hi @LaureenTh @geritwagner I am starting a set of issues here for the JOSE review

The checklist for the "Description" section of the paper has the following criterion:

> Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

I am not finding this in the current paper text, although I think I get the gist from the course itself.

If this is not in the paper already, can you add a one or two sentence summary of how you have used this course in the classroom, and how someone might adopt it?

For adoption, a simple statement should suffice, like "Someone interested in adopting our course can begin with the teaching notes" (and include a link to those notes, e.g. as a footnote)
