### 2025-08-11 digital-work-lab/open-source-project#237 @NickleDave
JOSE[paper]: Tell "story" of how authors came to develop course + expertise
The JOSE review checklist states:

> Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?

Passing comments in the paper indicate that parts of the course are based on previous work developing methods for teaching version control with Git. But I do not find a full description of how the course is developed. Please add a couple of sentences that speak to this.
