### 2025-08-11 digital-work-lab/open-source-project#238 @NickleDave
JOSE[doc]: add statement of need
The JOSE checklist asks for a statement of need in the documentation as well as the paper.

https://openjournals.readthedocs.io/en/jose/review_checklist.html#documentation

> A statement of need: Do the authors clearly state the need for this module and who the target audience is?

I think it could be sufficient to simply copy the statement of need from the paper to both the README and the summary page.

You might be repeating yourself a bit. But I am not finding a clear statement in the README / landing page of either the need, as it is outlined in the paper, or the target audience.

I would summarize your statement of need from the paper as "provide a full, structured course in Python package development". The repo and the documentation seem to be more generally about open source projects, with very little about Python packaging per se -- I'll address that in a separate issue. But if you want the docs to line up with the paper, you should at least state that in the docs.

And you should state in the docs:

> Our target audience is Bachelor students in Information Systems with initial programming experience.
