### 2025-08-11 digital-work-lab/open-source-project#239 @NickleDave
JOSE[doc]: add instructions for contributors, e.g. in CONTRIBUTING.md
The JOSE review checklist asks for:
https://openjournals.readthedocs.io/en/jose/review_checklist.html#documentation

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support?

I find plenty of links like "Improvement" but I don't find specific instructions like one would have in a [CONTRIBUTING.md](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file, e.g. we use just-the-docs + Jekyll, etc.

I would suggest
- [ ] add a CONTRIBUTING.md
- [ ] have all the tabs/buttons in the actual docs link to this page
- [ ] then have links *from* CONTRIBUTING.md to *templated* GitHub issues for specific contribution types, e.g. "feature" request / bugs / etc

but you definitely don't have to follow my suggestion. Just add more detail on how to contribute 🙂
