### 2025-08-11 #292 @NickleDave
Almost done with mine, one more issue to add but I don't have time to finish writing up today
