### 2025-08-12 fs-ise/open-source-project#241 @NickleDave
JOSE[pedagogy]: Make content on Python packaging substantial and complete
Hi again @LaureenTh @geritwagner

I really appreciate what you are trying to do with this course. I have tried to review with an understanding of how you see it.
But I feel that the content on Python packages does not meet several items on the JOSE review checklist.
Crucially I feel the content is not substantial or complete, and is not appropriate for the target audience, the second checklist item [here](https://openjournals.readthedocs.io/en/jose/review_checklist.html#pedagogy-instructional-design-work-in-progress-reviewers-please-comment).

My main issue is: the JOSE paper claims that the course provides a complete, structured course that teaches Python package development, targeted to 1st year bachelor's students in the Information Sciences. And it says that, by doing so, this course meets a critical need that is not addressed by resources like [py-pkgs.org](https://py-pkgs.org/welcome), that are instead targeted to motivated self-learners.

However:
- There is only [one section](https://digital-work-lab.github.io/open-source-project/docs/week_3_python_packages.html) on Python packaging, that is a sub-section of "week 3, Python". This makes it sound like you spend roughly one day out of one week on Python packaging.
- The content in the Python sections I would describe as (1) somewhat cursory and (2) not well written for the target audience

I will give examples below that support my second point about the content. If needed, I can break things down into separate issues.

I do not want this issue to be a wall of text that frustrates you. I do see you have put a tremendous amount of work into the course already. Please know I don't want to be an adversarial reviewer, and my intention is to help you get closer to achieving the goals you have laid out for the course.

At the highest level, I feel that, for the content to truly guide first year students in Information Sciences through Python packaging, you need highly simplified and concrete worked examples.

This approach of building up a Python package from scratch is taken by almost every resource on Python packaging I'm aware of
- py-pkgs: https://py-pkgs.org/welcome
- py-rse: https://third-bit.com/py-rse/packaging.html -- making a Python package is the culmination of their project writing functions around analyzing text with Zipf's law
- Vicki Boykis' Alice in Python Projectland: https://vickiboykis.com/2017/09/26/alice-in-python-projectland/
- Intermediate Software Development, from Software Carpentry: https://carpentries-incubator.github.io/python-intermediate-development/43-software-release.html

If these resources spend multiple consecutive sections on building up Python packages from scratch, and as you say they may be more targeted to motivated self-learners, then I think you would need *at least* as much content to guide your target audience through Python packaging.

If it were me, and I really wanted to teach students open source development including Python packaging, I would:
- [ ] rewrite the course to build a bare minimum version of ColRev from scratch.
  - [ ] that do not require the students to know or even think about anything that might confuse them because "magic happens" and/or contribute to cognitive overload, such as the real ColRev cli or GitHub codespaces
- [ ]  In other words, I would use a series of worked examples
    - [ ] that build upon core concepts from ColRev that you introduce in narrative text about the *goals* and *design* of the package -- I am thinking of a [domain-driven design approach](https://blog.nicholdav.info/posts/2025-01-02-domain-driven-design/) here. I *do not* think you need to actually introduce students to the idea of a [domain model](https://www.cosmicpython.com/book/chapter_01_domain_model.html), but you need to motivate the package design: "the goal of ColRev is to work with references for a literature review. References are often structured in BibTex format, where a set of fields with specific names are mapped to a set of values. We will build up a minimal version of ColRev from scratch that lets us work with references from bibtex ..."
    - [ ] So, just like the existing resources I linked to above, you should motivate each step with the goal of building the package
      - [ ] but you should introduce these with the minimum amount of jargon and without contributing to cognitive overload by using functionality from your actual package where "magic happens", see more specific examples below

I can understand why you would not want to restructure the entire course. At a bare minimum, to be able to tick off checkboxes for the review, I would think that you should at least flesh out the Python packaging section to give complete descriptions of everything, and provide worked examples. I'll give two examples that I encountered on my first couple of read-throughs:

1. As an example of how the material is a bit cursory (as I mentioned above):
  - you have a section called "How do I modify data items?" [here](https://digital-work-lab.github.io/open-source-project/docs/week_3_python_notebook_1.html#part-ii-data-items-)
  - where you describe Python dictionaries as "efficient data structures, which can be used to handle bibliographic records, such as the following (in BibTeX format)".
  - But you say nothing about how a dictionary *works*, you don't show any kind of worked example of *using* a dictionary. All you do is give a student a link to WS3, which I would say might even do more harm then good.
  - What is your [theory](https://cekrem.github.io/posts/programming-as-theory-building-naur/) of reference management that you use to design ColRev? How does that inform your decision to use dictionaries instead of, say, a dataclass?

For me, a bare minimum improvement on this would
- [ ] explicitly state the core concepts of using a dictionary -- this can be brief, since we assume the student has some experience with object-oriented languages. You can make this more interesting by giving some context here like "we use dictionaries in ColRev because the core functionality of the package involves frequent updates to records, and we need an efficient data structure like a dictionary that allows for that. Dictionaries map keys to values like so..."
- [ ] at least provide one worked example of using a dictionary to achieve some data transformation you do in ColRev, e.g., update the author, and then ask the student to complete a challenge using the dictionary, e.g., write a function that gets the author with graceful handling of the edge case where there is no author (which would be a good reason to introduce the `Dictionary.get()` method that returns a `None`)

2. As an example of why this is not well written for the target audience:
  - you give only a one paragraph description of the structure of a Python package
    - in py-pkgs, for example, the structure of a Package is an entire chapter ([this one](https://py-pkgs.org/04-package-structure))
  - in that paragraph you say "To create a Python package, it is essential to follow a clear structure that complies with best practices, such as those outlined in [PEP 517](https://peps.python.org/pep-0517/) and [PEP 518](https://peps.python.org/pep-0518/)."
    - even motivated self-learners may not know or care about what a PEP is. I think it's a common mistake (that I have made) to feel like you are "helping" people by introducing them to the idea of a PEP and pointing them to it. But there is not any additional information in the actual text here. Why do we care about those PEPs? Can we instead explain the goals of the PEPs without even linking to them? "Python packaging has in the past few years adopted a standardized format that uses declarative configuration files [1^]" "[^1] as described in PEP517 + 518"
  - you then ask students to use a special function built into `ColRev` to set up a project structure [here](https://digital-work-lab.github.io/open-source-project/docs/week_3_python_packages.html#setting-up-the-package). You say nothing about what this function does. When they go to develop their own package, they are not going to have access to a function like this.
 - I understand that you are trying to give the students hand on experience with open source contributions, but I think that the approach of providing a modified version of the package really adds to the cognitive load -- there's a lot of "call this function and magic happens"

A bare minimum rewrite that addresses these issues would at least
- [ ] fully explain the structure of a Python package, as is done here in the pyOpenSci docs: https://www.pyopensci.org/python-package-guide/tutorials/create-python-package.html#about-the-basic-package-directory-structure. It's not sufficient to just show the output of `tree`, you need to explain what each item is
- [ ] if you are going to suggest students use `uv`, then show them how to use `uv init` to set up a project from scratch
- [ ] don't teach students to use `colrev --init` and `colrev --check`, since they won't have this for their own packages, and you don't explain what these functions actually do
- [ ] remove/revise writing that is vague. I know I sound like grumpy graybeard reviewer number 2 here -- I am happy to raise more specific issues but I find many instances of writing that confuse me even though I supposedly know how to make a Python package and contribute to open source. Examples:
  - [ ] the section "Advanced testing strategy for ColRev" means nothing to me, and I have been writing unit tests for Python for 10 years. Why do first year information students need to know about your package's testing strategies?
  - [ ] Likewise you use the term "data  directory". You say "When using a package, it’s important to distinguish between the data directory and the package directory." -- I really don't know what this means. I think you are talking about a project a user has where there's a `./data` directory in the project root? Maybe consider adding a glossary? Are you trying to show the use of `pip install -e .` here? If so, why not write that instead?

We can discuss what I am asking for further on this issue so we do not add noise to the review.

Again, I realize GitHub Issues may not be the best medium to convey these ideas, and I don't want you to feel like you're reading a missive. I'd be happy to discuss on a Zoom meeting or something like that if it would be easier.

Thank you again for the work you are doing developing this course. I hope I can actually contribute to it.
