### 2025-08-14 085ebdf
add pdf command to Makefile
