### 2025-08-14 digital-work-lab/open-source-project#241 @NickleDave
Thank you @geritwagner -- I hope the feedback is mostly constructive and not too challenging 😅. I really appreciate your hearing me out. I definitely don't want this to be onerous. Maybe we can break things down into issues that meet the criteria and other issues for future enhancements to the course. Happy to work with you and @LaureenTh in whatever way makes the review most effective.
