### 2025-08-14 c963578
mention how the course has been used/can be adopted
closes #236
