+++ paper/paper.bib
@@ -79,7 +79,7 @@
 }
 
 @article{WagnerThurner2025,
-   author                        = {Wagner, G. and Thurner, L.},
+   author                        = {Wagner, Gerit and Thurner, L.},
    journal                       = {Journal of Information Systems Education},
    title                         = {Rethinking How We Teach Git: Pedagogical Recommendations and Practical Strategies for the Information Systems Curriculum},
    year                          = {2025},
