+++ docs/week_1_topic.md
@@ -17,12 +17,12 @@ nav_order: 2
 ![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
 
 In week 1, we start with the introductory session, covering the project outline and goals, explaining how groups will form and how contributions will be evaluated.
-We will start exploring the context of open-source projects and discuss any questions that may come up.
+We will start exploring the context of open source projects and discuss any questions that may come up.
 You do not need to register for the session or project in advance.
 We look forward to meeting you!
 
 {: .objective }
-> In this session, our goal is to familiarize ourselves with open-source projects on GitHub.
+> In this session, our goal is to familiarize ourselves with open source projects on GitHub.
 
 <br>
 
