### 2025-08-26 #292 @ctb
> [@ctb](https://github.com/ctb) thank you for completing your review checklist. Please let us know if you have any other feedback beyond the issues already raised in [@NickleDave](https://github.com/NickleDave) 's review.

I'm watching the issues, and will re-review once those are addressed.
