### 2025-09-14 15302ad
Update hall_of_fame.md
