+++ .wordlist.txt
@@ -992,3 +992,5 @@ fd
 jise
 jose
 theoj
+Chaoui
+ChloeT
