### 2025-11-03 0be0e16
Python packaging: revision plans
