### 2025-11-06 4cb3426
revise notebook-1
