+++ docs/week_3_python_notebook_1.md
@@ -34,9 +34,6 @@ With this notebook, you can familiarize yourself with Python syntax, create and 
 
 **"How do I write and use Python code?"**
 
-### Switch branch
-
-Navigate to the [CoLRev repository](https://github.com/CoLRev-Environment/colrev){: target="_blank"}, select the `tutorial_python` branch and start Codespaces.
 
 As a first step, we install the package dependency manager **uv**, which will be used in [part 3](#libraries):
 
@@ -47,28 +44,25 @@ pip install uv
 Next, we reset the state of the repository to the beginning of the tutorial:
 
 ```python
-git reset --hard 8fa941d167503d20659ea87ec6169affbf6574c3
+git reset --hard f0a13be2494181d56eb71a2f1aee8b4511f65919
 ```
 
 - As the session progresses, you can checkout the current commits
 - Whenever you see a `git reset --hard ...` command on the following slides, you can use it to set your repository to the required state (commit).
 
-### Setting up entrypoints
+### Setting up a Python script
 
-We implement a simple version of CoLRev that should be available through a separate command:
+We implement a simple Pyhton script, `run.py`, which should be available through the following command:
 
-```python
-colrev run
+```
+python run.py
 ```
 
-{: .info }
-> The previous command will initially create a `ModuleNotFoundError`. We will create this module in the next step.
+**Task**: Create the script, have it print "Hello World" and test the `python run.py` call.
 
-**Tasks:**
-
-- Check the last commit and the changes that were introduced. Which function does our new `run` command call?
--  Create the `run.py` module (module: file containing Python code) and the function that should be called. The function should print `Start simple colrev run`. Note that calling `colrev.ops.run.main()` means that colrev will try to import and run the `main()` function in the `/workspaces/colrev/colrev/ops/run.py` module.
-- Check the other functions in the module `/workspaces/colrev/colrev/ui_cli/cli.py`, and other modules in the `/workspaces/colrev/colrev` directory if necessary.
+```python
+git reset --hard 1778473e75718277ad2a1b623cde99c4fff674c0
+```
 
 ## Part II: Data items <a id="data"></a>
 
@@ -87,7 +81,7 @@ In this part, we focus on the data structure of dictionaries, which are often us
 }
 ```
 
-**Task**: Create a dictionary containing these data fields and print it when `colrev run` is called.
+**Task**: Create a dictionary containing these data fields and print it when `python run.py` is called.
 
 You can find the syntax for Python dictionaries (and many other data types) in the [W3School](https://www.w3schools.com/python/python_dictionaries.asp){: target="_blank"}.
 
@@ -108,7 +102,7 @@ Start simple colrev run
 To checkout the **solution**, run:
 
 ```python
-git reset --hard dbd583c499da9fe0945ce75f98c32c7babaddd4d
+git reset --hard 728a2dbe5a3c0f15e989eac4faab7b877b2f3a0c
 ```
 
 ## Part III: External libraries <a id="libraries"></a>
@@ -125,43 +119,21 @@ We decide to use the [BibtexParser](https://github.com/sciunto-org/python-bibtex
 pip install bibtexparser
 ```
 
-To add it as a dependency of CoLRev and make it available for users of the CoLRev package, we run
-
-
-```python
-uv add bibtexparser
-```
-
-<b>Task</b>: Check the changes and create a commit.
-
-To checkout the **solution**, run:
-
-
-```python
-git reset --hard c9d89be988d9ec8ab18b2841af74996bf2ad5b43
-cd /workspaces/colrev
-pip install -e .[dev]
-```
-
 ### Using external libraries
 
[…]
