### 2025-11-06 bb6df9a
update notebooks
