+++ docs/week_3_python_notebook_3.md
@@ -1,0 +1,12 @@
+---
+layout: default
+title: "Python notebook 3"
+parent: "Week 3: Python"
+nav_order: 4
+has_toc: true
+---
+
+# Exercise notebook: Python 3 (packages)
+
+
+cover pre-commit to the package tutorial part
