### 2025-11-06 151750f
add notes
