+++ slides/04-python_3.md
@@ -1,0 +1,23 @@
+---
+marp: true
+header: 'Open Source Project: Introduction to Python'
+theme: ub-theme
+paginate: true
+footer: "<a href='https://github.com/digital-work-lab/open-source-project/issues/new?template=Blank+issue' target='_blank'>♻️</a> <a href='https://github.com/digital-work-lab/open-source-project/edit/main/slides/04-python_2.md' target='_blank'>🛠️</a>"
+---
+
+<!-- _class: centered -->
+
+# WI-Project: Open Source Project
+
+## Introduction to Python (3)
+
+Prof. Dr. Gerit Wagner
+Faculty Information Systems and Applied Computer Sciences
+Otto-Friedrich-Universität Bamberg
+
+![center width:150px](../assets/qr-open-source-project.png)
+
+---
+
+TODO
