### 2025-11-06 c6d2a4a
Add session 3 content for creating a Python package with materials and installation instructions
