### 2025-11-06 7f18014
Update week 3 Python session details
Added details for Session 3 on creating a Python package, including materials and links.
