### 2025-11-06 c619994
Create tutorial for Python package development
Added a comprehensive tutorial on creating a Python package using `uv`, including prerequisites, project structure, and testing.
