+++ docs/week_3_python.md
@@ -80,7 +80,7 @@ If you have any questions during or after the sessions, we're here to assist. We
 
 <br>
 
-<iframe src="../output/05-python_3.html" 
+<iframe src="../output/07-python_3.html" 
         style="width: 100%; aspect-ratio: 16 / 9; border: 1px solid #ccc; border-radius: 5px;" 
         allowfullscreen>
 </iframe>
