### 2025-11-06 8a0dc15
Revise Python package creation documentation
Updated the documentation for creating a Python package, including prerequisites, installation instructions, and a step-by-step guide. Enhanced the structure and clarity of the content.
