+++ docs/week_3_python_packages.md
@@ -6,223 +6,231 @@ nav_order: 4
 has_toc: true
 ---
 
-# Python packages
+# Creating Your First Python Package
 
 ![Offered by: Digital Work at Otto-Friedrich-Universität Bamberg](https://img.shields.io/badge/Offered%20by-%20Digital%20Work%20(Otto--Friedrich--Universit%C3%A4t%20Bamberg)-blue)
 ![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
 
 ```mermaid
-graph LR
-    A["Python Packaging"] --> B0[<a href="#python-packages">Basics</a>] 
-    A --> B1[<a href="#testing">Testing</a>]
-    A --> B2[<a href="#documentation">Documentation</a>]
-    A --> B3[<a href="#publishing">Publishing</a>]
+graph TD
+    A["<b>Goal:</b> Create a sharable Python tool"] --> B["<b>1.</b> Learn the Anatomy of a Package"];
+    B --> C["<b>2.</b> Initialize Project with <code>uv init</code>"];
+    C --> D["<b>3.</b> Create the Directory Structure"];
+    D --> E["<b>4.</b> Install in Editable Mode"];
+    E --> F["<b>5.</b> Add a Function and a Test"];
+    F --> G["<b>6.</b> Manage Dependencies"];
+    G --> H["<b>Result:</b> A foundational, working package"];
+
 
     style A fill:#a2a2e2,stroke:#000,stroke-width:1px,color:#000
-    style B1 fill:#d199f1,stroke:#000,stroke-width:1px,color:#000
-    style B2 fill:#83cfc9,stroke:#000,stroke-width:1px,color:#000
-    style B3 fill:#a5d3f2,stroke:#000,stroke-width:1px,color:#000
-    style B0 fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style B fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style C fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style D fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style E fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style F fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style G fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style H fill:#d199f1,stroke:#000,stroke-width:1px,color:#000
 ```
 
-## Basics
+## 1. Introduction: What is a Python Package?
 
-To create a Python package, it is essential to follow a clear structure that complies with best practices, such as those outlined in [PEP 517](https://peps.python.org/pep-0517/){: target="_blank"} and [PEP 518](https://peps.python.org/pep-0518/){: target="_blank"}.
-The structure ensures that the package is maintainable, pip-installable, and ready for distribution.
-Below is an example of a typical package structure for CoLRev (Python) packages:
+A Python **package** is a standardized way to bundle and distribute reusable code so that others (or your future self) can easily install and use it with a simple `pip install` command.
+
+In this module, you will learn the fundamentals by building a complete, working package from scratch. Our goal is to demystify the process and understand the **why** behind each step. We will build a simple but practical utility package that standardizes journal names in bibliographic data, which is a core task in literature review tools like `Colrev`. This will give you the foundational skills needed to later build your own `Colrev` plugins.
+
+## 1b. Prerequisites: Setting Up Your Tools
+
+Before we begin, we need to ensure you have the necessary command-line tools. These do not come with Python and must be installed separately.
+
+### Installing `uv`
+
[…]
