### 2025-11-06 676684b
Remove images for installation verification
Removed installation verification images from documentation.
