### 2025-11-09 027f812
add the story of how the course was developed
closes #237
