+++ paper/paper.bib
@@ -40,7 +40,7 @@
 }
 
 @incollection{Malan2024,
-   doi                           = {10.1145/3649217.36535670},
+   doi                           = {10.1145/3649217.3653567},
    author                        = {Malan, David J},
    booktitle                     = {Proceedings of the 2024 on Innovation and Technology in Computer Science Education V. 1},
    title                         = {Containerizing CS50: Standardizing Students' Programming Environments},
@@ -95,4 +95,28 @@
   year      = {2020},
   isbn      = {9781138332250},
   url  = {https://py-pkgs.org/},
+}
+
+
+@article{EckhardtEtAl2025,
+  author       = {Eckhardt, Peter and Ernst, Katharina and Fleischmann, Thomas and Geßler, Anna and Schnickmann, Karl and Thurner, Laureen and Wagner, Gerit},
+  title        = {search-query: An Open-Source Python Library for Academic Search Queries},
+  journal      = {Journal of Open Source Software},
+  year         = {2025},
+  note         = {Under review},
+  institution  = {Otto-Friedrich Universität Bamberg},
+  url          = {https://github.com/openjournals/joss-reviews/issues/8775}
+}
+
+
+@article{Wagner2024,
+  author       = {Wagner, Gerit},
+  title        = {bib-dedupe: An open-source Python library for deduplication of bibliographic records},
+  journal      = {Journal of Open Source Software},
+  volume       = {9},
+  number       = {97},
+  pages        = {6318},
+  year         = {2024},
+  doi          = {10.21105/joss.06318},
+  url          = {https://github.com/CoLRev-Environment/bib-dedupe}
 }
