+++ .wordlist.txt
@@ -994,3 +994,8 @@ jose
 theoj
 Chaoui
 ChloeT
+MkDocs
+lifecycle
+declaratively
+backend
+distributable
