### 2025-11-09 92da8a5
Update week_3_python_packages.md
