+++ docs/week_3_python.md
@@ -31,12 +31,6 @@ If you have any questions during or after the sessions, we're here to assist. We
 > - Understand how to use other Python modules and libraries
 > - Prepare for extending the functionality of the CoLRev package
 
-> [!NOTE]  
-> Plans for the revision:
-> This page will be based on the [colrev-python-tutorial](https://github.com/CoLRev-Environment/colrev-python-tutorial) repository.
-> Students will run the code outside of a Python package, using `python script.py` (instead of colrev run).
-> For simplicity, data and source code will be in the same directory.
-
 ## Session 1
 
 <br>
@@ -86,7 +80,7 @@ If you have any questions during or after the sessions, we're here to assist. We
 
 <br>
 
-<iframe src="../output/05-python_3.html" 
+<iframe src="../output/07-python_3.html" 
         style="width: 100%; aspect-ratio: 16 / 9; border: 1px solid #ccc; border-radius: 5px;" 
         allowfullscreen>
 </iframe>
@@ -102,31 +96,6 @@ After you have used CoLRev in the last notebook, it is instructive to create a p
 To do this, we recommend starting a Codespace from the fork of your group.
 You will also need to create a separate project directory in the Codespace (as explained in the notebooks).
 Try to run the operations presented in the [asciinema recording](https://colrev-environment.github.io/colrev/){: target="_blank"} in the CoLRev documentation and analyze how the data in the project (Git) repository changes.
-
----
-
-## Session 3
-
-> [!NOTE]  
-> Plans for the revision:
-> Session 3 will be dedicated to Python packaging. Students will work in the [colrev-python-package-tutorial](https://github.com/CoLRev-Environment/colrev-python-package-tutorial) and create a colrev package "from scratch". Here, we should introduce the distinction between data and src directories (by running pip install ..., the package can be called without specifying a path).
-> TODO: Convert the Python packaging page to a presentation and create a corresponding practice notebook.
-
-<br>
-
-<iframe src="../output/04-python_3.html" 
-        style="width: 100%; aspect-ratio: 16 / 9; border: 1px solid #ccc; border-radius: 5px;" 
-        allowfullscreen>
-</iframe>
-
-<br>
-
-| **Materials**       | **Link**                                                                                                  |
-|----------------------|---------------------------------------------------------------------------------------------------------|
-| Python session 3     | [![View Slides Part 3](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/04-python_3.html){: target="_blank"} [![View PDF Slides](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/04-python_3.pdf){: target="_blank"} |
-| Notebook session 3   | [![Open Notebook](https://img.shields.io/badge/Notebook-blue?logo=github)](week_3_python_notebook_3.html) |
-
----
 
 {: .deliverable }
 ![tasks logo](../assets/iconmonstr-clipboard-5.svg)Contribute code to a fork, create branches and merge code ([deadline](../index.html#deliverables)).
