### 2025-11-19 ab02f60
add CONTRIBUTING.md
closes #239
