### 2025-11-19 a314be8
add CONTRIBUTING.md
closes #239
