+++ CONTRIBUTING.md
@@ -1,0 +1,89 @@
+# Contributing to *The Open Source Project*
+
+Thank you for your interest in contributing to *The Open Source Project*! We welcome contributions of all kinds -- from bug fixes and content improvements to new examples or simply questions and feedback. This document outlines how you can:
+
+- **Contribute improvements or new content** to the module
+- **Report issues or problems** you encounter
+- **Seek support or ask questions** about using the module
+
+By following these guidelines, you help us maintain a high-quality, user-friendly learning resource for everyone.
+
+## Project Overview and Setup
+
+This project is a learning module built with [Jekyll](https://jekyllrb.com/) using the *Just-the-Docs* theme and managed via GitHub Pages and workflows. All course content (including slides and notebooks) is generated from Markdown files in this repository. **Contributors do not need to compile the site manually** -- if your contribution is merged, our GitHub Actions will automatically build and update the site, slides, and notebooks. However, if you wish to preview your changes locally, you can run the site with Jekyll (see Jekyll's documentation for setup) or use GitHub Codespaces.
+
+**Repository structure in brief:** Most of the instructional content resides under the `docs/` directory (for narrative content and pages) and the `slides/` directory (for lecture slides). Changes to these Markdown files will be reflected in the website and slide deck after our automated build. There are also configuration files (e.g., `_config.yml`) and workflows that manage things like spelling checks and link checks. When contributing, you typically will be editing the Markdown content or related resources.
+
+## How to Contribute Improvements 🚀
+
+We are grateful for any form of improvement. Here are common ways to contribute:
+
+- **Fixing typos or minor issues:** Feel free to directly submit a pull request for small fixes like typos, broken links, or formatting issues. These are quick to review and merge.
+- **Improving documentation or examples:** If you can clarify an explanation, add an example, or expand a section for better understanding, please do! You can propose the change via a pull request.
+- **Adding new content or features:** For larger contributions (e.g., a new exercise, a new section, or a significant change in structure), it\'s usually best to start by opening an issue to discuss the idea before you invest a lot of effort. This allows us to ensure the change fits with our course layout and instructional design.
+- **Translations or adaptations:** If you want to translate content or adapt materials for another context, please open an issue to discuss how best to incorporate these contributions.
+
+When contributing content, **please ensure consistency with the module's layout and instructional design**. Maintain the style and tone of existing materials -- for example, follow the formatting of the syllabus and weekly outlines, use similar heading levels, and keep code examples in the same format as current ones. Consistency helps learners have a seamless experience. If you're unsure about the style, feel free to ask for guidance in an issue.
+
+### Submitting a Pull Request (PR)
+
+If you have a specific improvement ready (code or content), you can submit it directly as a pull request:
+
+1.  **Fork the repository** to your GitHub account and then clone your fork to your local machine.
+2.  **Create a new branch** for your changes. Give it a descriptive name (e.g., `fix-typo-intro-section`).
+3.  **Make your changes** in the appropriate files. Most often, this will be editing or adding Markdown files in the `docs/` or `slides/` folders. If you're contributing code (e.g., in notebooks or scripts), ensure it follows our coding standards (we use Python coding best practices, and our repository includes pre-commit hooks for formatting and linting).
+4.  **Test your changes** if possible. Also, check that the Markdown formatting is correct and any links you added aren't broken.
+5.  **Commit your changes** with a clear commit message explaining *what* you changed and *why*. For example: `"Fix typo in Week 3 Python notebook description"` or `"Add example on packaging to Week 5 section"`.
+6.  **Push the branch** to your fork on GitHub.
+7.  **Open a pull request** from your branch into our `main` branch. In the PR description, clearly describe the change you are proposing. If the PR addresses an open issue, reference the issue number (e.g., "Closes #12") so that the linkage is clear.
+
[…]
