+++ _config.yml
@@ -46,4 +46,4 @@ aux_links:
   "Repository":
     - "https://github.com/digital-work-lab/open-source-project"
   "Improvement":
-    - "https://github.com/digital-work-lab/open-source-project/issues"
+    - "https://github.com/digital-work-lab/open-source-project/blob/main/CONTRIBUTING.md"
