### 2025-11-19 digital-work-lab/open-source-project#239 @geritwagner
Thank you for this helpful suggestion. We concur that a CONTRIBUTING.md is valuable for signaling to students and external users how they can contribute to the project. We added a CONTRIBUTING.md file with the following:

- Outlines how contributors can suggest changes to markdown-based materials (site content, slides, and notebooks), all of which are built using just-the-docs and Jekyll workflows.
- Provides guidance on whether to open an issue or a pull request, and clarifies how to contribute in line with the module's instructional design.
- Includes links to the issue tracker and templates for bugs, feature requests, and questions.
- Explicitly acknowledges the role of course evaluations: feedback collected from students at the end of each semester is integrated into the issue tracker, allowing us to transparently trace how we address suggestions through specific changes over time.

We also updated the navigational structure of the documentation (e.g. “Improvement” links) to point directly to the CONTRIBUTING.md page.
