### 2025-11-19 fb8d8c9
link CONTRIBUTING.md in README
