### 2025-11-19 925b980
Update paper.md
