### 2025-11-19 72e6dd5
add statement of need to REAMDE and docs/index
close #238
