### 2025-11-19 #292 @geritwagner
Thanks for checking in, @ashleefv ! We have addressed a few of the issues and are currently working on the more substantial (and clearly important) revisions of [241](https://github.com/digital-work-lab/open-source-project/issues/241). Our plan is to push a complete revision by the end of November so we can continue the discussion with Reviewer 1 on [241](https://github.com/digital-work-lab/open-source-project/issues/241).
