### 2025-11-20 e49b04e
Python packaging revision  (#242)
* Python packaging: revision plans

* revise notebook-1

* update notebooks

* add notes

* Add session 3 content for creating a Python package with materials and installation instructions

* move changes to appropriate branch

* add CONTRIBUTING.md

closes #239

---------

Co-authored-by: Emediong Francis <emediongfrancis@gmail.com>
Co-authored-by: gerit.wagner@uni-bamberg.de <you@example.com>
