+++ .wordlist.txt
@@ -1010,3 +1010,4 @@ Scalability
 modularize
 reusability
 scalability
+env
