+++ docs/week_5_best_practices.md
@@ -25,7 +25,7 @@ The session also includes practical examples and guidelines that will help you d
 
 <br>
 
-<iframe src="../output/05-best_practice.html" 
+<iframe src="../output/06-best_practice.html" 
         style="width: 100%; aspect-ratio: 16 / 9; border: 1px solid #ccc; border-radius: 5px;" 
         allowfullscreen>
 </iframe>
@@ -34,7 +34,7 @@ The session also includes practical examples and guidelines that will help you d
 
 | **Materials**                | **Link**                                                                                                                                  |
 |------------------------|-----------------------------------------------------------------------------------------------------------------------------------------|
-| Slides  | [![View Slides Part 1](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/05-best_practice.html){: target="_blank"} [![View PDF Slides](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/05-best_practice.pdf){: target="_blank"} |
+| Slides  | [![View Slides Part 1](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/06-best_practice.html){: target="_blank"} [![View PDF Slides](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/06-best_practice.pdf){: target="_blank"} |
 
 {: .info }
 > **Checklist**
