+++ docs/week_3_python_notebook_3.md
@@ -8,9 +8,228 @@ has_toc: true
 
 # Exercise notebook: Python 3 (packages)
 
-contents:
+![Offered by: Digital Work at Otto-Friedrich-Universität Bamberg](https://img.shields.io/badge/Offered%20by-%20Digital%20Work%20(Otto--Friedrich--Universit%C3%A4t%20Bamberg)-blue)
+![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
 
-- package setup (pyproject.toml)
-- unit tests
-- pre-commit to the package tutorial part
-- plugin for colrev (pyproject.toml, base-classes, colrev env --update-package-list)
+```mermaid
+graph TD
+    A["<b>Goal:</b> Create a shareable Python tool"] --> B["<b>1.</b> Learn the Anatomy of a Package"];
+    B --> C["<b>2.</b> Initialize Project with <code>uv init</code>"];
+    C --> D["<b>3.</b> Create the Directory Structure"];
+    D --> E["<b>4.</b> Install in Editable Mode"];
+    E --> F["<b>5.</b> Add a Function and a Test"];
+    F --> G["<b>6.</b> Manage Dependencies"];
+    G --> H["<b>Result:</b> A foundational, working package"];
+
+
+    style A fill:#a2a2e2,stroke:#000,stroke-width:1px,color:#000
+    style B fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style C fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style D fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style E fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style F fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style G fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
+    style H fill:#d199f1,stroke:#000,stroke-width:1px,color:#000
+```
+
+## 1. Introduction: What is a Python Package?
+
+A Python **package** is a standardized way to bundle and distribute reusable code so that others (or your future self) can easily install and use it with a simple `pip install` command.
+
+In this module, you will learn the fundamentals by building a complete, working package from scratch. Our goal is to demystify the process and understand the **why** behind each step. We will build a simple but practical utility package that standardizes journal names in bibliographic data, which is a core task in literature review tools like `Colrev`. This will give you the foundational skills needed to later build your own `Colrev` plugins.
+
+## 1b. Prerequisites: Setting Up Your Tools
+
+Before we begin, we need to ensure you have the necessary command-line tools. These do not come with Python and must be installed separately.
+
+### Installing `uv`
+
+`uv` is a modern, extremely fast tool for managing Python packages and projects. We will use it to initialize our project and handle dependencies.
+
+Open your terminal or command prompt and run the following command:
+
+```bash
+# For macOS, Linux, and Windows WSL
+curl -LsSf https://astral.sh/uv/install.sh | sh
+
+# For Windows (Powershell)
+irm https://astral.sh/uv/install.ps1 | iex
+```
+After installation, close and reopen your terminal. Verify it was installed correctly by running:
+
+```bash
+uv --version
+```
+
+You should see the installed version number printed.
+
+### Installing `pytest`
+
+`pytest` is the framework we will use to write and run tests for our code. While we could install it globally like `uv`, it's a best practice to install testing tools as **development dependencies** for each project. We will do this in Step 3.5.
+
+## 2. The Anatomy of a Modern Python Package
+
[…]
