### 2025-11-26 64732de
update Python notebooks
