+++ docs/week_3_python_notebook_3.md
@@ -11,34 +11,33 @@ has_toc: true
 ![Offered by: Digital Work at Otto-Friedrich-Universität Bamberg](https://img.shields.io/badge/Offered%20by-%20Digital%20Work%20(Otto--Friedrich--Universit%C3%A4t%20Bamberg)-blue)
 ![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
 
-```mermaid
-graph TD
-    A["<b>Goal:</b> Create a shareable Python tool"] --> B["<b>1.</b> Learn the Anatomy of a Package"];
-    B --> C["<b>2.</b> Initialize Project with <code>uv init</code>"];
-    C --> D["<b>3.</b> Create the Directory Structure"];
-    D --> E["<b>4.</b> Install in Editable Mode"];
-    E --> F["<b>5.</b> Add a Function and a Test"];
-    F --> G["<b>6.</b> Manage Dependencies"];
-    G --> H["<b>Result:</b> A foundational, working package"];
+We  <img src="../assets/iconmonstr-favorite-2.svg" alt="Edit" width="12" height="12">  your <a href="https://github.com/digital-work-lab/open-source-project/issues/new/choose" target="_blank">feedback</a> and <a href="https://github.com/digital-work-lab/open-source-project/edit/main/docs/week_3_python_notebook_2.md" target="_blank">suggestions</a> on this notebook!
 
+---
 
-    style A fill:#a2a2e2,stroke:#000,stroke-width:1px,color:#000
-    style B fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
-    style C fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
-    style D fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
-    style E fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
-    style F fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
-    style G fill:#f2d08b,stroke:#000,stroke-width:1px,color:#000
-    style H fill:#d199f1,stroke:#000,stroke-width:1px,color:#000
-```
+This notebook guides you through creating a small but fully functional Python package. Step by step, you will set up a clean project structure, write reusable code, add automated tests, and manage dependencies using modern tooling. By the end, you will have a shareable package that follows current best practices—and a solid foundation for developing more advanced tools in the future. Specifically, we will cover the following learning objectives:
 
-## 1. Introduction: What is a Python package?
+{: .objective } 
+> - Learn how to set up a Python package using uv
+> - Extend CoLRev by making the package available as a plugin
+
+| Part        | Label                                                                      | Time (min) |
+|-------------|----------------------------------------------------------------------------|------------|
+|             | [Setup: Tools & environment](#prerequisites-setting-up-your-tools)         |     10     |
+|  Part VIII  | [Creating a package](#part-viii-creating-a-package)                        |    110     |
+|  Part IX    | [CoLRev plugin context](#part-ix-colrev-plugin-context)                    |    40      |
+|             | [Conclusion and further steps](#5-conclusion-and-further-steps)            |    20      |
+|             | Overall                                                                    |    180     |
+
+## Setup
+
+**What is a Python package?**
 
 A Python **package** is a standardized way to bundle and distribute reusable code so that others (or your future self) can easily install and use it with a simple `pip install ...` command.
 
 In this module, you will learn the fundamentals by building a complete, working package from scratch. Our goal is to demystify the process and understand the **why** behind each step. We will build a simple but practical utility package that standardizes journal names in bibliographic data, which is a core task in literature review tools like `colrev`. In the last part, we will learn how to make the new package available as a plugin to `colrev`.
 
[…]
