+++ docs/week_3_python_notebook_3.md
@@ -6,7 +6,7 @@ nav_order: 4
 has_toc: true
 ---
 
-# Exercise notebook: Python 3 (packages)
+# Exercise notebook: Python 3
 
 ![Offered by: Digital Work at Otto-Friedrich-Universität Bamberg](https://img.shields.io/badge/Offered%20by-%20Digital%20Work%20(Otto--Friedrich--Universit%C3%A4t%20Bamberg)-blue)
 ![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
