### 2025-11-28 987eb26
Enhance week 3 notebook with detailed explanations on `colrev` base classes and code quality practices. Introduce the concept of base classes for plugin development and the role of `pre-commit` hooks in maintaining code quality.
