### 2025-11-30 33c0295
revise package development notebook
