+++ docs/week_3_python_notebook_3.md
@@ -15,19 +15,23 @@ We  <img src="../assets/iconmonstr-favorite-2.svg" alt="Edit" width="12" height=
 
 ---
 
-This notebook guides you through creating a small but fully functional Python package. Step by step, you will set up a clean project structure, write reusable code, add automated tests, and manage dependencies using modern tooling. By the end, you will have a shareable package that follows current best practices—and a solid foundation for developing more advanced tools in the future. Specifically, we will cover the following learning objectives:
+This notebook guides you through creating a small but fully functional Python package and turning it into a plugin of CoLRev. Step by step, you will set up a clean project structure, write reusable code, add automated tests, and manage dependencies using modern tooling. By the end, you will have a shareable package that follows current best practices—and a solid foundation for developing more advanced tools in the future. Specifically, we will cover the following learning objectives:
 
 {: .objective } 
 > - Learn how to set up a Python package using uv
 > - Extend CoLRev by making the package available as a plugin
 
-| Part        | Label                                                                      | Time (min) |
-|-------------|----------------------------------------------------------------------------|------------|
-|             | [Setup](#setup)                                                            |     10     |
-|  Part VIII  | [Creating a package](#part-viii-creating-a-package)                        |    110     |
-|  Part IX    | [CoLRev plugin context](#part-ix-colrev-plugin-context)                    |    40      |
-|             | [Conclusion and further steps](#5-conclusion-and-further-steps)            |    20      |
-|             | Overall                                                                    |    180     |
+| Part   | Label                                      | Time (min) |
+|--------|--------------------------------------------|------------|
+|        | Setup                                      | 10         |
+| Part 1 | Initialization                             | 20         |
+| Part 2 | Structure & Code                           | 20         |
+| Part 3 | Editable Install                           | 10         |
+| Part 4 | Testing                                    | 20         |
+| Part 5 | Code Quality                               | 20         |
+| Part 6 | Base Classes & CoLRev Plugin Context       | 70         |
+|        | Wrap-up                                    | 10         |
+|        | **Overall**                                | **180**    |
 
 ## Setup
 
@@ -59,9 +63,11 @@ uv --version
 
 You should see the installed version number printed.
 
-## The anatomy of a modern Python package
+## Part 1: Initialization
 
+Let's start with the anatomy of a Python package.
 A consistent structure is key. It allows automated tools and other developers to understand your project instantly.
+A typical package could be structured as follows:
 
 ```
 colrev-journal-formatter/
@@ -79,17 +85,18 @@ colrev-journal-formatter/
 *   `tests/`: The **testing folder**. Contains code that automatically checks if your source code works correctly.
 *   `README.md` & `LICENSE`: Your project's documentation and legal rules.
 
-## Part VIII: Creating a package
-
 Let's build a package named `colrev-journal-formatter`.
 
-### Step 1: Initialize the project with `uv`
-
-First, create a directory and initialize the project.
+First, create a directory and navigate into it.
 
 ```bash
 mkdir colrev-journal-formatter
 cd colrev-journal-formatter
+```
+
+Now, we use `uv` to initialize the project. This automatically creates the `pyproject.toml` file, which is the control center for modern Python packages.
+
+```bash
 uv init
 ```
 
@@ -115,7 +122,14 @@ build-backend = "hatchling.build"
 *   **`dependencies = []`**: A list of other packages that your package needs to function. `uv add` will populate this list for you.
[…]
