### 2025-11-30 7406387
update session overviews
