### 2025-11-30 bbcd0ca
move slides to session pages
