### 2025-11-30 bae1410
revisions/consistency
