### 2025-12-01 194d470
revise python notebooks
