### 2025-12-01 7205f81
update slides links
