+++ docs/week_3_python_notebook_1.md
@@ -24,7 +24,8 @@ With this notebook, you can familiarize yourself with Python syntax, create and 
 > - Learn how to use external libraries
 > - Learn how to use functions
 
-Complementary session slides are available here: [![View Slides Part 1](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/03-python_1.html){: target="_blank"} [![View PDF Slides](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/03-python_1.pdf){: target="_blank"}
+📑 Session Slides
+➡️ [![View online](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/03-python_1.html){: target="_blank"} · ⬇️ [![Download PDF](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/03-python_1.pdf){: target="_blank"}
 
 | Part       | Label                              | Time (min) |
 |------------|------------------------------------|------------|
