+++ docs/week_3_python_notebook_3.md
@@ -21,7 +21,8 @@ This notebook guides you through creating a small but fully functional Python pa
 > - Learn how to set up a Python package using uv
 > - Extend CoLRev by making the package available as a plugin
 
-Complementary session slides are available here: [![View Slides Part 3](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/05-python_3.html){: target="_blank"} [![View PDF Slides](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/05-python_3.pdf){: target="_blank"}
+📑 Session Slides
+➡️ [![View online](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/05-python_3.html){: target="_blank"} · ⬇️ [![Download PDF](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/05-python_3.pdf){: target="_blank"}
 
 | Part   | Label                                      | Time (min) |
 |--------|--------------------------------------------|------------|
