+++ docs/week_3_python_notebook_2.md
@@ -23,7 +23,8 @@ With this notebook, you can refactor to object-oriented code, switch to a real C
 > - Learn how to save and commit changes
 
 📑 Session Slides
-➡️ [![View online](https://img.shields.io/badge/View-Slides-orange?logo=html5)](../output/04-python_2.html){: target="_blank"} · ⬇️ [![Download PDF](https://img.shields.io/badge/Download-PDF-orange?logo=adobe)](../output/04-python_2.pdf){: target="_blank"}
+
+➡️ [View online](../output/04-python_2.html){: target="_blank"} · ⬇️ [Download PDF](../output/04-python_2.pdf){: target="_blank"}
 
 
 | Part      | Label                                                        | Time (min) |
