+++ docs/week_3_python_notebook_3.md
@@ -253,33 +253,49 @@ A very common framework for this is called **`pre-commit`**. It works by setting
 *   Check for common programming errors or bugs (linting).
 *   Ensure documentation is up to date.
 
-TODO: add example pre-commit hook and install
+Create a `.pre-commit-config.yaml` configuration file with the following content, which adds the [black](https://github.com/psf/black) code formatter:
 
-In a project that uses this tool (like the main `colrev` repository or the repository for this course itself), a developer would use the following command to manually run all the checks:
+```yaml
+repos:
+  - repo: https://github.com/psf/black
+    rev: 24.3.0
+    hooks:
+      - id: black
+        language_version: python3
+```
+
+Use the following command to install and run all the checks:
 
 ```bash
-# This is an example of a command run in a mature, configured project.
+# Install the pre-commit framework (a Python package)
+pip install pre-commit
+
+# Install pre-commit into the Git repository, ensuring that they run with every git commit ... 
+pre-commit install
+
+# Run the pre-commit hooks
 pre-commit run --all
 ```
 
 If any of the checks fail, the commit is blocked until the developer fixes the issues. This acts as an automated gatekeeper, ensuring high quality for all code added to the project.
 
-<!-- 
-**You do not need to run this for your small tutorial package**, as setting up a `pre-commit` system is an advanced topic. However, it is a key part of the modern development workflow, and you will see it in almost any established open-source project you contribute to in the future.
- -->
-
 **Breakpoint:** Jump to the solution for Step 5.
 
-TODO
 ```bash
-git reset --hard XXXXXX
+git reset --hard 6566a670a8888ef3177232e8e4b9b322c2529afd
+pip install pre-commit
+pre-commit install
 ```
 
 ## Part 6: CoLRev plugins <a id="plugins"></a>
 
 To turn our package into a CoLRev plugin, we need to inherit from the `colrev` Base Classes and specify the entry-point in our `pyproject.toml`.
 
-**TODO: install colrev !**
+Start by installing colrev:
+
+```bash
+pip install colrev
+```
 
 So far, we have written a standalone function, `standardize_journal_name`. This is great for a simple package, but how does a function like this become a true, integrated part of a large application like `colrev`? The answer is through **base classes**.
 
@@ -297,6 +313,7 @@ To illustrate, here is what that would look like conceptually. **You do not need
 # A conceptual example of a real colrev plugin.
 
 import colrev.package_manager.package_base_classes as base_classes
+import colrev.record.record
 
 # Our new class inherits from the 'PrepPackageBaseClass' base class, fulfilling the contract.
 class JournalAbbreviationPrep(base_classes.PrepPackageBaseClass):
@@ -305,7 +322,7 @@ class JournalAbbreviationPrep(base_classes.PrepPackageBaseClass):
         super().__init__(prep_operation, settings)
 
     # By implementing the 'prepare' method, our class can be called by colrev.
-    def prepare(self, record: colrev.record.Record) -> colrev.record.Record:
+    def prepare(self, record: colrev.record.record.Record) -> colrev.record.record.Record:
         """Standardizes journal names in a CoLRev record"""
 
         if 'journal' in record.data:
@@ -336,9 +353,8 @@ By inheriting from `PrepPackageBaseClass` and placing our logic inside the `prep
 
 **Breakpoint:** Jump to the solution.
 
-TODO
 ```bash
-git reset --hard b4bf67e8de8026d2f7dd845c5c459211e938e107
+git reset --hard 044c1d985b0a14df6c8842afefe6027f3579a0f7
 ```
 
 As a last step, we need to specify how `colrev` will interact with our package.
[…]
