+++ docs/week_3_python_notebook_3.md
@@ -15,11 +15,11 @@ We  <img src="../assets/iconmonstr-favorite-2.svg" alt="Edit" width="12" height=
 
 ---
 
-This notebook guides you through creating a small but fully functional Python package and turning it into a plugin of CoLRev. Step by step, you will set up a clean project structure, write reusable code, add automated tests, and manage dependencies using modern tooling. By the end, you will have a shareable package that follows current best practices—and a solid foundation for developing more advanced tools in the future. Specifically, we will cover the following learning objectives:
+This notebook guides you through creating a small but fully functional Python package and turning it into a plugin of `colrev`. Step by step, you will set up a clean project structure, write reusable code, add automated tests, and manage dependencies using modern tooling. By the end, you will have a shareable package that follows current best practices—and a solid foundation for developing more advanced tools in the future. Specifically, we will cover the following learning objectives:
 
 {: .objective } 
 > - Learn how to set up a Python package using uv
-> - Extend CoLRev by making the package available as a plugin
+> - Extend `colrev` by making the package available as a plugin
 
 <br>
 
@@ -35,9 +35,9 @@ This notebook guides you through creating a small but fully functional Python pa
 | Part 3 | [Editable installation](#editable_install) | 10         |
 | Part 4 | [Testing](#testing)                        | 20         |
 | Part 5 | [Code quality](#code_quality)              | 20         |
-| Part 6 | [CoLRev plugins](#plugins)                 | 70         |
+| Part 6 | [colrev plugins](#plugins)                 | 70         |
 |        | Wrap-up                                    | 10         |
-|        | **Overall**                                | **180**    |
+|        | **Overall*-                              | **180*-  |
 
 {: .info }
 > Start a Codespace in the [colrev-python-package tutorial](https://github.com/CoLRev-Environment/colrev-python-package-tutorial){: target="_blank"} repository.
@@ -95,10 +95,10 @@ colrev-journal-formatter/
 └── pyproject.toml
 ```
 
-*   `pyproject.toml`: The **control center**. It contains the package name, version, dependencies, and build instructions.
-*   `colrev_journal_formatter/`: The **source folder**. Your actual Python code lives here.
-*   `tests/`: The **testing folder**. Contains code that automatically checks if your source code works correctly.
-*   `README.md` & `LICENSE`: Your project's documentation and legal rules.
+- `pyproject.toml`: The **control center**. It contains the package name, version, dependencies, and build instructions.
+- `colrev_journal_formatter/`: The **source folder**. Your actual Python code lives here.
+- `tests/`: The **testing folder**. Contains code that automatically checks if your source code works correctly.
+- `README.md` & `LICENSE`: Your project's documentation and legal rules.
 
 Let's build a package named `colrev-journal-formatter`.
 
@@ -133,11 +133,11 @@ build-backend = "hatchling.build"
 
 Adapt the sections as you see fit (e.g., update the `authors` and add the `build-system`).
 
-*   **`[project]`**: A standard section holding all your project's metadata.
-*   **`name`**: The name that users specify to install your package (`pip install colrev-journal-formatter`).
-*   **`version`**: The current version. You should increment this with each new release.
-*   **`dependencies = []`**: A list of other packages that your package needs to function. `uv add` will populate this list for you.
-*   **`[build-system]`**: This section tells `pip` *how* to build your package. It specifies the "build backend" (`hatchling`) which acts as a "factory" to assemble your code into a distributable format.
+- **`[project]`**: A standard section holding all your project's metadata.
+- **`name`**: The name that users specify to install your package (`pip install colrev-journal-formatter`).
[…]
