### 2025-12-01 4aaca49
update audience and session figure
