+++ .wordlist.txt
@@ -1011,3 +1011,9 @@ modularize
 reusability
 scalability
 env
+pkgs
+rse
+Breakpoint
+Boykis
+Projectland
+formatter
