### 2025-12-02 92b07fc
spelling and consistency
