+++ slides/02-git.md
@@ -8,12 +8,12 @@ footer: "<a href='https://github.com/digital-work-lab/open-source-project/issues
 
 <!-- _class: centered -->
 
-# WI-Project: Open Source Project
+# WI-Project: Open source project
 
 ## Introduction to Git
 
-Prof. Dr. Gerit Wagner
-Faculty Information Systems and Applied Computer Sciences
+Prof. Dr. Gerit Wagner  
+Faculty of Information Systems and Applied Computer Sciences  
 Otto-Friedrich-Universität Bamberg
 
 ![center width:150px](../assets/qr-open-source-project.png)
@@ -23,8 +23,8 @@ Otto-Friedrich-Universität Bamberg
 # Check-in: Group formation
 
 - [Milestone](https://github.com/CoLRev-Environment/colrev/milestones)
-- Anyone not yet part of an issue discussion?
-- Challenges related to the setup?
+- Is anyone not yet part of an issue discussion?
+- Are there any challenges related to the setup?
 
 ![bg right width:500px](../assets/group-formation.png)
 
@@ -34,16 +34,16 @@ Otto-Friedrich-Universität Bamberg
 
 Advantages:
 
-- Every repository has a full version history
-- Most operations run locally
-- Reliable data handling, ensuring integrity and availability
-- Efficient data management for versions and branches
-- Scalable collaboration mechanisms for large teams and complex projects
+- Every repository has a full version history.
+- Most operations run locally.
+- Reliable data handling ensures integrity and availability.
+- Efficient data management for versions and branches.
+- Scalable collaboration mechanisms for large teams and complex projects.
 
 Caveats:
 
-- Need to learn and understand the underlying model
-- Not built for binary files or large media files
+- There is a need to learn and understand the underlying model.
+- Git is not built for binary files or large media files.
 
 ![bg right:35% width:300px](../assets/git_logo.png)
 
@@ -59,19 +59,19 @@ blockquote {
 
 # Learning objectives
 
-Understand and use git to develop software in teams.
+Understand and use Git to develop software in teams.
 
-**Part 1**: Branching
-**Part 2**: Committing
-**Part 3**: Collaborating
+**Part 1**: Branching  
+**Part 2**: Committing  
+**Part 3**: Collaborating  
 
 Each part starts with the **concepts** before the **practice** session.
 
 In the practice sessions:
 
-- Form groups of two to three students
-- Work through the exercises
-- Create a *cheat sheet* summarizing the key commands
+- Form groups of two to three students.
+- Work through the exercises.
+- Create a *cheat sheet* summarizing the key commands.
 
 ![bg right:45% width:620px](../assets/reorder.png)
 
@@ -105,11 +105,11 @@ blockquote {
 }
 </style>
 
-- A **commit** refers to a snapshot (version) of the whole project directory, including the metadata and files
-- The **tree** object contains all files (and non-empty directories); it is identified by a SHA hash
-- Commits are created in a sequence, with every commit pointing to its **parent** commit(s)
-- Commits are identified by the **SHA** fingerprint of their metadata and content\*, e.g., `98ca9`
-- Commits are created by the **git commit** command
+- A **commit** refers to a snapshot (version) of the whole project directory, including the metadata and files.
+- The **tree** object contains all files (and non-empty directories); it is identified by a SHA hash.
+- Commits are created in a sequence, with every commit pointing to its **parent** commit(s).
+- Commits are identified by the **SHA** fingerprint of their metadata and content\*, e.g., `98ca9`.
+- Commits are created by the **git commit** command.
 
 ![bg right:45% width:230px center](../assets/git-commit.png)
 
@@ -127,11 +127,11 @@ blockquote {
 
 # The DAG, branches, and HEAD
 
[…]
