+++ docs/hall_of_fame.md
@@ -5,11 +5,11 @@ has_toc: true
 nav_order: 11
 ---
 
-# Hall-of-fame
+# Hall of fame
 
 On this page, we celebrate the accomplishments of our students 🎉 🍾 🎈
 
-Student contributors to the Open Source project [CoLRev](https://github.com/CoLRev-Environment/colrev){: target="_blank"} with the following GitHub accounts:
+Student contributors to the Open Source Project [CoLRev](https://github.com/CoLRev-Environment/colrev){: target="_blank"} with the following GitHub accounts:
 
 - dengdenglele
 - MalouSchmidt
@@ -48,4 +48,4 @@ Student contributors to the Open Source project [CoLRev](https://github.com/CoLR
 - Lea-Chaoui
 - ChloeT17
 
-See complete overview of [contributors](https://github.com/CoLRev-Environment/colrev/graphs/contributors){: target="_blank"}.
+See a complete overview of [contributors](https://github.com/CoLRev-Environment/colrev/graphs/contributors){: target="_blank"}.
