### 2025-12-04 b15d488
update agenda figure in paper
