### 2026-02-09 #292 @geritwagner
Hi @ashleefv, thanks for checking in!

We have completed substantial revisions addressing the major points raised by Reviewer 1 (@NickleDave), particularly around the Python packaging content and course structure. These changes were informed by detailed feedback, and we’ve shared the revised materials with the reviewer to ensure they are aligned with his expectations.

At the moment, we’re waiting for follow-up feedback on the revised contents. Once we hear back, we plan consolidate everything quickly and complete the revisions.
