### 2026-02-09 #292 @ashleefv
@NickleDave have you had a chance to review the revised materials?
