### 2026-02-27 fs-ise/open-source-project#241 @geritwagner
Thanks you, @NickleDave — really appreciate it! 🙌 Looking forward to your thoughts, and thanks again for taking the time to re-review!
