### 2026-03-02 #292 @NickleDave
Hi @ashleefv so sorry I missed your ping -- I have now reviewed the revised materials.

As I let the authors know [here](https://github.com/fs-ise/open-source-project/issues/241#issuecomment-3987698667), I feel that they have addressed all the issues I raised.

I also updated the checklist above to show all items as complete.

I thank the authors for gracefully listening to my constructive criticism. I find the lessons and paper to be much improved.
