### 2026-03-03 #292 @geritwagner
Hi @ashleefv, thanks for the nudge!

And thanks again to @NickleDave — as highlighted in [#241](https://github.com/fs-ise/open-source-project/issues/241), we truly appreciate the careful re-review and confirmation that the issues have been addressed.

@ctb, we’re happy to move forward. Please let us know if there’s anything further we should take care of — we’re glad to address any remaining points.
