### 2026-03-05 #292 @geritwagner
Thank you very much. We have now completed the requested steps. In addition, we took the opportunity to do a final round of polishing and proofreading of the manuscript.

- Tagged software release: v0.3.0
- Archived version: Zenodo
- DOI: https://doi.org/10.5281/zenodo.18868452

The Zenodo record has been checked to ensure that the metadata matches the paper, including the correct title and author list. We also added the ORCIDs.

Please let us know if anything further is needed on our side. Thank you again for handling the submission and for the helpful reviews that improved the paper.
