### 2026-03-06 #292 @ashleefv
@geritwagner can you check the DOIs that were skipped or missing in the references? You can repeat this test by asking the editorialbot to "check references"
