### 2026-03-06 #292 @geritwagner
@ashleefv I have fixed the missing DOI and reviewed the skipped ones. The skipped DOIs are ok from my view. Two are books, and the others are a conference paper and a journal paper where no DOI was assigned.
