### 2026-03-06 #292 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
