+++ README.md
@@ -1,13 +1,13 @@
 <p align="center">
-    <img alt="Project Logo" src="https://raw.githubusercontent.com/digital-work-lab/open-source-project/main/assets/open-source-project.png" width="600px">
+    <img alt="Project Logo" src="https://raw.githubusercontent.com/fs-ise/open-source-project/main/assets/open-source-project.png" width="600px">
 </p>
 
 <div align="center">
 
-[![Offered by: Digital Work at Otto-Friedrich-Universität Bamberg](https://img.shields.io/badge/Offered%20by-%20Digital%20Work%20(Otto--Friedrich--Universit%C3%A4t%20Bamberg)-blue)](https://digital-work-lab.github.io/open-source-project/)
-[![Spelling Check](https://github.com/digital-work-lab/open-source-project/actions/workflows/spelling.yml/badge.svg)](https://github.com/digital-work-lab/open-source-project/actions)
-![Links Fix](https://github.com/digital-work-lab/open-source-project/actions/workflows/links_fix.yml/badge.svg)
-![Documentation Status](https://img.shields.io/github/actions/workflow/status/digital-work-lab/open-source-project/pages.yml?label=documentation)
+[![Offered by: Digital Work at Otto-Friedrich-Universität Bamberg](https://img.shields.io/badge/Offered%20by-%20Digital%20Work%20(Otto--Friedrich--Universit%C3%A4t%20Bamberg)-blue)](https://fs-ise.github.io/open-source-project/)
+[![Spelling Check](https://github.com/fs-ise/open-source-project/actions/workflows/spelling.yml/badge.svg)](https://github.com/fs-ise/open-source-project/actions)
+![Links Fix](https://github.com/fs-ise/open-source-project/actions/workflows/links_fix.yml/badge.svg)
+![Documentation Status](https://img.shields.io/github/actions/workflow/status/fs-ise/open-source-project/pages.yml?label=documentation)
 ![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-green.svg)
 [![DOI](https://zenodo.org/badge/697812022.svg)](https://doi.org/10.5281/zenodo.13871275)<br>
 [![JISE](https://img.shields.io/badge/JISE-10.62273%2FBTKM5634-green)](https://jise.org/Volume36/n1/JISE2025v36n1pp1-12.html)
@@ -21,9 +21,9 @@
 Welcome! The Open Source Project is organized by the digital work lab (Prof. Dr. Gerit Wagner) at Otto-Friedrich-Universität Bamberg.
 You can find more information on the following pages:
 
-- [Project page](https://digital-work-lab.github.io/open-source-project/)
-- [Syllabus](https://digital-work-lab.github.io/open-source-project/docs/syllabus.html)
-- [Teaching notes](https://digital-work-lab.github.io/open-source-project/docs/teaching_notes.html)
+- [Project page](https://fs-ise.github.io/open-source-project/)
+- [Syllabus](https://fs-ise.github.io/open-source-project/docs/syllabus.html)
+- [Teaching notes](https://fs-ise.github.io/open-source-project/docs/teaching_notes.html)
 
 ## Statement of need
 
@@ -101,11 +101,11 @@ Feedback and suggestions for improvement are always welcome (see our [CONTRIBUTI
 
 ## Reusing materials (instructors)
 
-To reuse the materials, you can copy individual elements or [fork](https://github.com/digital-work-lab/open-source-project/fork) the repository.
+To reuse the materials, you can copy individual elements or [fork](https://github.com/fs-ise/open-source-project/fork) the repository.
 
 ## Pedagogical Foundations
 
[…]
