+++ CITATION.cff
@@ -6,14 +6,17 @@ authors:
     orcid: "https://orcid.org/0000-0003-3926-7717"
   - family-names: "Thurner"
     given-names: "Laureen"
+    orcid: "https://orcid.org/0009-0000-5716-5816"
   - family-names: "Tang"
     given-names: "Carlo"
+    orcid: "https://orcid.org/0009-0009-3168-4859"
   - family-names: "Ott"
     given-names: "Stella"
+    orcid: "https://orcid.org/0009-0002-5191-1237"
 title: "fs-ise/open-source-project"
-version: 0.1.0
+version: 0.4.0
 identifiers:
   - type: doi
     value: 10.5281/zenodo.13871275
-date-released: 2024-10-04
+date-released: 2026-02-25
 url: "https://github.com/fs-ise/open-source-project?tab=readme-ov-file"
