### 2026-03-25 #292 @geritwagner
Thank you! I just went through the author checklist:

- [x] Authors and affiliations (including ORCIDs) double-checked
- [x] Software release: **v0.4.0**
- [x] Archive on Zenodo: **DOI: 10.5281/zenodo.19225336**
- [x] Archive metadata (title, authors, ORCIDs) matches the JOSE paper
- [x] License: **CC-BY-4.0** in the archive and repository.

Excited to see the paper "in print" 🎉
