+++ CITATION.cff
@@ -17,6 +17,6 @@ title: "fs-ise/open-source-project"
 version: 0.4.0
 identifiers:
   - type: doi
-    value: 10.5281/zenodo.13871275
+    value: 10.5281/zenodo.19225336
 date-released: 2026-02-25
 url: "https://github.com/fs-ise/open-source-project?tab=readme-ov-file"
