### 2026-03-26 a95ef40
Bump lycheeverse/lychee-action from 1 to 2.0.2 in /.github/workflows (#251)
Bumps [lycheeverse/lychee-action](https://github.com/lycheeverse/lychee-action) from 1 to 2.0.2.
- [Release notes](https://github.com/lycheeverse/lychee-action/releases)
- [Commits](https://github.com/lycheeverse/lychee-action/compare/v1...v2.0.2)

---
updated-dependencies:
- dependency-name: lycheeverse/lychee-action
  dependency-version: 2.0.2
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com
[…]
