+++ .github/workflows/links_fix.yml
@@ -19,7 +19,7 @@ jobs:
 
       - name: Link Checker
         id: lychee
-        uses: lycheeverse/lychee-action@v1
+        uses: lycheeverse/lychee-action@v2.0.2
 
       - name: Create or Update Link Checker Issue
         if: env.lychee_exit_code != 0
