### 2026-03-28 #292 @labarba
Reading through the paper proof now... I have a question: what are your sources for the enrollment numbers presented on Table 1?

For example, I pulled up "Programming for Everybody: Getting Started withPython" on Coursera and it shows "3,537,693 already enrolled"—the table shows 581,247: where is this from?
