### 2026-03-28 #292 @labarba
Ah! That's the [edX version of the course](https://www.edx.org/learn/python/the-university-of-michigan-programming-for-everybody-getting-started-with-python), now showing "595,065 learners enrolled"

Given that Chuck's courses are both in Coursera and edX, you might update the numbers in the table to reflect that?
