### 2026-03-28 #292 @labarba
The "Python for Everybody Specialization" shows enrollment of 212,000 on Table 1, but the [Coursera page](https://www.coursera.org/specializations/python/paidmedia) says "1,935,482 already enrolled."
