### 2026-03-28 #292 @labarba
"Python for Data Science, AI and Development" shows enrollment of 37,000 on Table 1, but the [Coursera page](https://www.coursera.org/learn/python-for-applied-data-science-ai) says "1,462,579 already enrolled"—so it looks like all these numbers need to be reviewed and updated in the paper, if you really want to display the enrollment column, that is. (Not sure how meaningful those "enrollment" numbers are.)
