### 2026-03-28 #292 @labarba
And since you will do a touch-up of the paper, go ahead and review the references, applying caps protection with `{}` when needed; e.g., "bloom" "python" (twice)
