### 2026-03-28 #292 @geritwagner
Thank you for the careful check.

You’re right that the enrollment numbers are inconsistent across platforms and change over time. I’ve removed the enrollment column from Table 1, as it’s not essential and would quickly become outdated.

I also reviewed the references and added capitalization protection where needed. Happy to revise other parts as well if you spot anything that would benefit from clarification or polishing.

Appreciate the helpful feedback!
