### 2026-03-28 #292 @labarba
Congratulations, @geritwagner – your JOSE paper is published! 🚀

Huge thanks to our Editor: @ashleefv and Reviewers: @NickleDave, @ctb  – we couldn't do this without you 🙏
