### 2026-03-28 #292 @geritwagner
Huge thanks to the entire review panel for your thoughtful feedback and support throughout the process! 🙏

I’m excited to see the paper published and very grateful for the effort everyone invested. Special appreciation as well to the editors for guiding the process so smoothly.

And I’m especially proud of @LaureenTh, @dengdenglele, and @Stella1234-design—wonderful teamwork all the way through!
